refactor(enginenetx): make stats API private
Now that we've more or less reached the point where we wanted to be with
https://github.com/ooni/probe/issues/2531, let's spend some time to
refactor the implementation, now that we know very well how to works,
such that modifying it in the future would be easier.

The first order of business here seems to hide implementation details
and get rid of too many HTTPSDialer prefixes which only cause confusion
when looking at the available structs.
